chore: split FixtureRunner from fixtures
It is only used in worker process.
